Split client module to entities
