+++ tests/client/test_utils.py
@@ -4,8 +4,8 @@ from codecs import encode
 
 import mock
 
-from scrapinghub.utils import parse_auth
-from scrapinghub.utils import format_iter_filters
+from scrapinghub.client.utils import parse_auth
+from scrapinghub.client.utils import format_iter_filters
 
 
 def test_format_iter_filters():
